+++ admin/class-tw-chat-admin.php
@@ -244,6 +244,8 @@
             $chat_widget_name = sanitize_text_field($_POST['tw_chat_widget_name']);
             $greeting = sanitize_text_field($_POST['tw_chat_greeting']);
             $suggested_answers = sanitize_text_field($_POST['tw_chat_suggested_answers']);
+            $dismiss_answers = sanitize_text_Field($_POST['tw_chat_dismiss_answers']);
+            $dismiss_answers_text = sanitize_text_Field($_POST['tw_chat_dismiss_answers_text']);
             $assistant_id = sanitize_text_field($_POST['tw_chat_assistant_id']);
             $email_recipients = sanitize_text_field($_POST['tw_chat_email_recipients']);
             $webhook_address = sanitize_text_field($_POST['tw_chat_webhook_address']);
@@ -271,6 +273,8 @@
             update_post_meta($post_id, 'tw_chat_assistant_id', $assistant_id);
             update_post_meta($post_id, 'tw_chat_greeting', $greeting);
             update_post_meta($post_id, 'tw_chat_suggested_answers', $suggested_answers);
+            update_post_meta($post_id, 'tw_chat_dismiss_answers', $dismiss_answers);
+            update_post_meta($post_id, 'tw_chat_dismiss_answers_text', $dismiss_answers_text);
             update_post_meta($post_id, 'tw_chat_webhook_address', $webhook_address);
             update_post_meta($post_id, 'tw_chat_webhook_header', $webhook_header);
             update_post_meta($post_id, 'tw_chat_email_recipients', $email_recipients);