+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Project Changelog
 
+## 1.2.2
+- Added ListInput component for better management of the suggest answers field in the chat widget admin form.
+- Added widget fields to allow users to clear suggested answers.
+- Fixed log file caching in the Log File Viewer.
+- Restructure admin UI page components for readability and performance.
+
 ## 1.2.1
 - Links in the chat response now open in a new tab.
 - Added token usage to plugin log.